Se trabajó en la implementación de imegenes y relaciones entre entidades

+++ src/main/java/com/example/manosalaobrabackend/repository/ComentarioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ComentarioRepository extends JpaRepository<Comentario, Integer> {
+public interface ComentarioRepository extends JpaRepository<Comentario, Long> {
 }
